Add Arithmetic trait for CPU Register
