+++ crates/cpu/src/lib.rs
@@ -9,6 +9,7 @@ pub mod nextpc;
 mod opcodes;
 pub mod reader;
 pub mod registers;
+pub(crate) mod arithmetic;
 
 pub use crate::bus::RegisterBus;
 pub use crate::cpu::Cpu;
@@ -16,3 +17,4 @@ pub(crate) use crate::futures::{GetAt, SetAt};
 pub use crate::interface::{NewRegisters, Registers};
 pub(crate) use crate::reader::Reader;
 pub(crate) use flags::Flags;
+pub(crate) use crate::arithmetic::Arithmetic;
